backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
pytesseract==0.3.10
httpx==0.26.0
reportlab==4.0.9
gunicorn==21.2.0
PyJWT==2.8.0
flask-limiter==3.5.0

